Very early implementation of LHS rendering
I'd like to be able to write entries in Literate Haskell style.  This is
a hacky series of additions which allow that to happen.  As yet, it's
not complete, and is a reasonably ugly solution.

Still, it works, and I can start refactoring on it later.  The reality
is that this blog is supposed to be a platform for my writing, and so
long as it's standing up to that I'm not *too* fussed by the relative
badness.